给系统设置独立权限来登录后台；删除标签、系统判断加前提条件（关联已经没有）等
git-svn-id: https://svn.chsi.com.cn/svn/repos/chsicommunity/knowledge@17993 5607bd42-195c-0410-82fb-b7440d194b30

+++ trunk/webapp/src/main/java/com/chsi/knowledge/htgl/action/KnowledgeAction.java
@@ -265,10 +265,14 @@ public class KnowledgeAction extends AjaxAction {
     public void delKnowledge() throws Exception {
         if (!ValidatorUtil.isNull(id)) {
             KnowledgeData data = knowledgeService.getKnowledgeById(id);
-            knowIndexService.deleteKnowIndex(data.getId());// 删索引
-            CmsServiceClient cmsServiceClient = CmsServiceClientFactory.getCmsServiceClient();
-            cmsServiceClient.deleteArticle(data.getCmsId());// 从新闻系统删除
-            knowledgeService.delete(data);// 从本系统删除
+            if(data!=null) {
+                knowIndexService.deleteKnowIndex(data.getId());// 删索引
+                CmsServiceClient cmsServiceClient = CmsServiceClientFactory.getCmsServiceClient();
+                cmsServiceClient.deleteArticle(data.getCmsId());// 从新闻系统删除
+                data.setKnowledgeStatus(KnowledgeStatus.YSC);
+                knowledgeService.update(data);// 从本系统删除,非真删除
+                knowTagRelationService.del(data.getId());
+            }
             ajaxMessage.setFlag(Constants.AJAX_FLAG_SUCCESS);
         } else {
             ajaxMessage.addMessage("id不能为空！");
